Fix repo finding and storing being depended on nilable value
The download install link is nullable, so in case the link is nil we're now getting the available download link and using that instead. This fixes the issue of some repos only being in a testing phase crashing the program because we're trying to access something on a nil memory address.